Bug#32192454: WL14340: RPDSERVER SIG11 AT RPDMPR_BVFLT_RPDMPR_DTYPE_SB8_RPDOQC_OPT_TYPE_GT_TRUE
The metadata of an expression using the UNHEX function can be wrong.
If the argument of the UNHEX function is a LONGTEXT column, the
calculation of UNHEX's maximum length overflows a 32-bit integer in an
intermediate result and becomes zero.

Fixed by performing the calculation with 64-bit precision so that the
intermediate result doesn't overflow.

Change-Id: I8165f1344b3875e151a4abd350b65b59f1611610