Bug#32211954: PROVIDE HYPERGRAPH AND JOIN TO SECONDARY ENGINE COST HOOK
When a secondary engine estimates the cost of an access path, it may
need information in the JoinHypergraph object or the JOIN object, for
example to examine the predicates associated with the access path.

Pass a reference to the JoinHypergraph object as an argument to the
secondary engine cost hook, and make the Query_block and JOIN object
accessible via JoinHypergraph::query_block() and
JoinHypergraph::join().

Change-Id: Ic014328176fa5dcfed743d28cca1f22e5a719dc5