Bug #32215939 MOCK SERVER TAKES LONG TIME TO RESPOND TO EACH QUERY.
On a slower machine with one core I noticed that it takes up to 0.5 second
for the mysql_server_mock to handle SINGLE query statement. That makes
some of the test very slow and causes some failures (limited timeouts).

The problem is that most of our tracefiles call the prepare_statements
function inside a function that handles a query. It means that it is
called for every single query each session executes. prepare_statements
is quite heavy and there is no reason to repeat it for each query.

This patch changes our mock server tracefiles, moving the part that is
preparing the statements to the global scope - which means it is executed
only once per each connection/session.

RB: 25577