Bug#32223182 http_server test hangs for 20sec
Problem
=======

routertest_component_http_server randomly hands for 20seconds and then
fails on slow machines.

Starting the http-server and creating the passwd files that it depends
on happens in parallel.

If the passwd file is created before http-server is started, the test
passes.

If the http-server starts before the passwd is written, the test fails.

As the start of the http-server is not synchronized with sd-notify, the
test may speak the server listening socket before the signal handler is
installed, and lead to a hang.

Change
======

- create passwd file before starting http_server
- use sd-notify socket to check for readiness
- changed testnames to underscores in code to allow easier lookup

RB: 25587