Bug #32224279 ROUTERTEST_COMPONENT_ASYNC_REPLICASET OCCASIONALLY THROWS AN EXCEPTION
The following exception sometimes can be seen when running
routertest_component_async_replicaset

[ RUN      ] AsyncReplicasetTest.OnlyPrimaryLeftAcceptsRW
unknown file: Failure
C++ exception with description "Error connecting to MySQL
server at 127.0.0.1:11015: Lost connection to MySQL server
at 'reading initial communication packet', system error: 0
(2013)" thrown in the test body.
/home/areliga/dev/server/mysql-8.0/router/tests/helpers
/process_manager.cc:484 : Failure

Because of the exception there are no logs nor console
outputs from any of the processes and it is hard to
investigate why the test is failing.

This patch changes the tests to use functions that handle
the exceptions when making the client connections.

RB: 25590