Bug #32227090 TEST_REST_ROUTING::ROUTING_STATUS_123 FAILS RANDOMLY ON SLOWER MACHINES
This failure is seen quite often on slowe machine when running tests in
parallel:

test_rest_routing.cc:250: Failure
Expected equality of these values:
  value->GetInt() Which is: 0
  expected_blocked_hosts Which is: 1

The problem is that the test is causing the blocking of a client then
sleeps for 200ms and hopes this is enough for the Router to process
this blocking. This is not true for slower machine.

This patch waits for specific entry in the Router log instead of
sleeping for a fixed period.

RB: 25595