Bug #31546136 NDB_RESTORE_CONV_REMAP_DEBUG FAILS DUE TO BLOB BACKUP LOG ORDER
ndb_restore_conv_remap_debug fails with a 'Corrupted blob value'
error. This error occurs intermittently while replaying backup
logs which delete blobs values.

A blob value is stored with a small fixed-size part in the main
table which contains the blob column, and the remaining data
in longer parts in an NDB$PARTS table which is specific to that
blob column. To delete a blob value, there are separate backup logs
for the main table row and the parts.

The problem is that the main table row DELETE uses the synchronous
API, which triggers DELETEs for all the blob parts. So the subsequent
part DELETES always fail with 'Not found' errors. These errors are
ignored since ndb_restore ignores 'Not found' errors for DELETEs.

The use of the sync API can result either in correct data or in
corrupted data, depending upon the ordering of the logs:

1) Delete main table row -> delete part rows
   - Main table row delete deletes all parts
   - Part row deletes fail silently
2) Delete parts -> delete main table row
   - Part row deletes delete all parts
   - Main table row delete fails silently as part rows are already missing
   - Blob has corrupted main table row without parts, error on reads

Both log orders are valid, since there is no guaranteed ordering
in the backup log.

This is fixed by modifying ndb_restore to use the async API for
blob deletes. The async API does not trigger blob part deletes
when a blob main table row is deleted. So the main table delete
logevent deletes only the main table row.

This fix does not implement logentry batching, so serialised log
replay remains the same.

Reviewed by: Frazer Clement <frazer.clement@oracle.com>

Change-Id: Ie08cce3cfd74086f6b93bd79aca35e47b3d23b87