WL#14474 ADDS SUPPORT TO USE FASTCOV TOOL AND REFACTORS ATRT SCRIPTS
Modifies atrt to accepts a new command line option --coverage-tool,
which can take two values "lcov"(default) or "fastcov". Modifies atrt's
analyze and compute coverage script to accept arguments in an arbitary
order, refactors all variable names to be in lower case and adds a
function to display script usage information. Autotest version has been
updated from 10 to 11.

This newly added option decides the tool that is gonna be used for
coverage analysis.

autotest-run.sh is modified to accept this command line option and pass
the same to atrt.

This option is of type ENUM and throws error is invalid values are
passed to the option.

Information regarding the tool to be used is passed from atrt to
atrt-analyze-coverage.sh and atrt-compute-coverage.sh scripts.
Commands to compute coverage with these tools have been changes
accordingly.

Change-Id: Ifabc5c4f2802255d82507c877f1e710ff3a7f332