BUG#32045786 NO ERROR MESSAGE WHEN MGMD IS STARTED ON PORT THAT IS ALREADY IN USE
When a management node is started on a port that is already in use, the
error is written into the logfile. But no error message is displayed on
console from where the node is being started. Users should have to look
into logs to find out why the node is failing to start.

This fix adds validation in ConfigRetriever::verifyConfig method that
binds to hostname and portnumber specified in the config for mgmd node.
If binding fails error is thrown to the console indicating that address
is already in use.

Change-Id: Ieadcb01b4cbba53b3a9b2ee61a0a70fa3395dd84