Bug#32217164 prepare_statement_responses is slow
Problem
=======

prepare_statement_responses() of the mock-server's common_statements
module prepares a statements the mock-server is expected to understand
for this test.

It takes the stmt-keys and prepares the response for each statement.

For each statement

* it calls common_statements.get()
* common_statements.get() prepares all statements and returns one response.

which means that for each statement the same statements are prepared
each run even though responses don't change.

As a side-effect of the speedup of the mock-server,
test_gr_notifications started to fail as more notifications were
delivered than the test expected.

Change
======

- prepare responses once for all stmt_keys in prepare_statment_responses()
- in wait_md_queries() wait until _at least_ min_queries are seen.

| routertest_component        | Before  | After   |
| --------------------------- | ------- | ------- |
| mock_server                 |  2.72 s |  2.77 s |
| routing_connection          | 26.77 s | 24.50 s |
| routing_splicer             | 33.62 s | 32.42 s |
| bootstrap_account           | 36.44 s | 33.11 s |
| metadata_http_auth_backend  | 57.37 s | 43.55 s |
| rest_routing                | 51.39 s | 49.60 s |
| rest_api_enable             | 49.65 s | 49.89 s |
| metadata_ttl                | 73.77 s | 61.44 s |
| gr_notification             | 77.51 s | 36.03 s |

RB: 25579