Bug #32232769 METADATA_TTL AND ROUTING_STRATEGY TESTS DEPEND ON UNRELIABLE SLEEPS
test_metadata_ttl and test_routing_strategy still are using
sleep with fixed timeout to wait for some initialization to occur.
This is unreliable for slow machines.

This patch changes the code to:
1) use the notify socket for discovering if the Router is ready
to accept connections
2) use the information from mock server to check for the Router
to run metadata refresh

Additionally the patch fixes the printing issue of the millisecond
parameters when failed.

RB: 25606