Bug#32180383 NDB_INIT() CRASHES WHEN CHECKING L3 CACHE ON ARM64
Ndb_ReloadHWInfo() is responsible for updating Hardware info for all the
CPUs in the machine. Linux ARM does not have L3 cache information. So, For
Linux ARM, NdbHW module assigns socket id for L3 cache id. But, this function
failed to record the value for important global variable num_shared_l3_caches,
Which is required when creating lists of CPUs connected to a same L3 cache.
As num_shared_l3_caches points to default NdbHW aborts the process.

Fixed the issue by recording the value for num_shared_l3_caches

Change-Id: I99a8d7188d4619162a63c2a3d9d44bd0a5f17fda