Bug#32227101: CREATE TABLE PRODUCES CORE
Problem: Specifying a very large value (exeeding UINTMAX32) for
key_block_size when creating a table would lead
to an assert.

The value is represented as uint32 internally, but parser
allowed an ulong.

Solution: As the largest allowed value is 16, we change the grammar rule
for this number to num_int (as there is no num_uint), so that value
causing overflow are rejected.

Change-Id: I95cd65bd03e53889ff04369372f878ac5178b068