Bug#32247424 SHOW CLUSTER CONFIG GENERATION
Make it possible to query the generation of the configuration used by the
NdbApi in the MySQL Server.

When the MySQL Server connects to the cluster, the NdbApi will fetch the
cluster configuration and then use it to configure connections and other
parameters of the NdbApi. There is a generation number in the
configuration which can be used as an indicator to determine if the
configuration of the cluster has changed since the NdbApi was configured.

Add new global status variable "Ndb_config_generation". Also extend the
NdbApi with new method for extracting the generation of
the currently used config.

  mysql> SHOW GLOBAL STATUS LIKE 'Ndb_config_generation';
  +-----------------------+-------+
  | Variable_name         | Value |
  +-----------------------+-------+
  | Ndb_config_generation | 37    |
  +-----------------------+-------+
  1 row in set (0.01 sec)

  mysql> SELECT VARIABLE_VALUE FROM performance_schema.global_status
      -> WHERE VARIABLE_NAME = 'Ndb_config_generation';
  +----------------+
  | VARIABLE_VALUE |
  +----------------+
  | 37             |
  +----------------+
  1 row in set (0.04 sec)

Change-Id: I7a605e4964466123013dbadeb48b320ed553d430