Bug#32175578 TEST FAILED STARTUP SHUTDOWN IN ASAN
Followup fix.

The ndb_startup and ndb_connection_pool_nodeids
tests fail on Windows with error 1 when testing
the abnormal startup. This is caused by redundant
semicolon on the mysqld command line which are
interpreted as illegal arguments on Windows while
it's ignored on other platforms.

This error can be spotted in the mysqld error log:
  Too many arguments (first extra is ';').

Fix by removing the redundant semicolon at end of
mysqld command line.

Change-Id: Id108fde94b6791c3e0b787ce9b53d304d172e3c5