BUG#32210176 MGMT NODE SHOULD NOT START WITH INVALID NODEGROUP NUMBER
When Management node is started with config with invalid Nodegroup for
Data nodes, no error was thrown. And ndbd fails to start with that
config. Maximum Nodegroups allowed is 72.

Config rule check_node_vs_replicas was previously checking if Nodegroup
is greater that MAX_NDB_NODES which is now modified to check against
MAX_NDB_NODE_GROUPS and throw error when Invalid Nodegroup is specified.

Change-Id: Ie3c27c781378c1d27101de63bc87730221116f6d