Bug#32231557: Field_blob::store_to_mem: assertion `field_charset == cs' failed.
The problem here is that we have a mismatch between the character set
assigned to an Item_func_nullif object and the target string of its
result evaluation. The reason for this is that
Item_func_nullif::resolve_type_inner() aggregates the comparison
collation into the target collation (member collation) instead of
the comparison collation (member cmp.cmp_collation). The evaluation
later takes collation from the first argument to NULLIF, which has
become different from the item's collation.

The fix is to use the correct member in the comparison type aggregation.

Reviewed by: Knut Anders Hatlen <knut.hatlen@oracle.com>