BUG#32258207 MGMD SHOULD NOT START WITH REPLICAS=1 AND DATANODES >72
When number of replicas is 1, the number of allowed datanodes is equal
to number of nodegroups allowed. But Mgmd was not failing when more
than 72 Datanodes were added in Config file.

Config rule "check_node_vs_replicas" is modified by adding validation
that checks if number of datanodes is less than Max nodegroups allowed.

Change-Id: I27d10292bf90a1b9288ff4087bb70a7e12ea0ff2