Bug #31684783 	MEB.PARALLEL_UNDO_TRUNCATE FAILURE [part 1/3]
On WinAPI the os_file_delete_if_exists_func() gets the same trick that nt_share_delete() does:
Deleting a file may fail if some other process uses it. However, the file might have been opened with FILE_SHARE_DELETE mode, in which case the delete will succeed, but the file will not be deleted, only marked for deletion when all handles are closed. To work around this, we first move the file to new randomized name, which will not collide with a real name.
The copy of the code in os_file_delete_func() was removed.
Sleep before failure is reduced from 2000x1s to 20x0.1s.
The meb.parallel_undo_truncate test in tweaked to force undo tablespace truncatation and cause MySQL server failure on Release version on Windows before the fix.
Added a testcase to stress undo truncation while opening the tablespace file with and without FILE_SHARE_DELETE flag.

RB#25557