Bug#32299045: Crash after JOIN::generate_derived_keys, concurrent triggers
Problem here is that we access the THD pointer for use with getting
the mem_root pointer for allocating memory in add_derived_key(),
as well as in calls to actual_key_parts() and actual_key_flags().
This happens because we access TABLE::in_use from internal temporary
tables which are associated with triggers. Those triggers have been
transferred from one session to another, but the corresponding in_use
pointers have not been updated.

The fix is to replace use of TABLE::in_use with current_thd.

Reviewed by: Chandan Kunal <chandan.kunal@oracle.com>