Bug#32248316: ENABLE THE HYPERGRAPH OPTIMIZER FOR INSERT INTO ... SELECT [2/2, enable]
Enable use of the hypergraph optimizer for INSERT INTO ... SELECT
statements and REPLACE INTO ... SELECT statements.

The only case that needed some special care was INSERT SELECT
statements with an ON DUPLICATE KEY UPDATE clause, and where the
topmost access path was reading from a temporary table. In this case,
the items representing the values in the ON DUPLICATE KEY UPDATE
clause needed to be rewritten to point to fields in the temporary
table.

Change-Id: Ifdb4f99232a8e2b89facef8ab8b1b64279313342