Bug#32255538: LONG LOG WRITING MIGHT BLOCK LOG FLUSHING OPPORTUNITY WHEN INNODB-LOG-WRITER-THREADS=OFF
"os_event_reset(log.old_flush_event)" works as exclusive lock for log_flush for innodb-log-writer-threads=off mode.
And not better to hold it when log_writer_write_buffer() at 8.0.
Because the log_writer_write_buffer() might take long time for the loop or waiting for log file free/archiving or etc...
The other threads might be able to finish with the current write_lsn flushed. They might be blocked if exists.

So, fix points for 8.0 are
- "os_event_reset(log.old_flush_event)" better to be after "log_writer_write_buffer()" called enough.
- attempts to release writer_mutex temporally for the other threads' opportunity.
  (for each several MB written. (log_write_ahead_size[several KB] * 1024 should be enough))

RB: 25633
Reviewed-by: Pawel Olchawa <pawel.olchawa@oracle.com>