WL#14474 HANDLES ATRT AND AUTOTEST-RUN.SH ARGUMENTS UNIFORMLY
Modifies autotest-run.sh script to uniformly handle arguments passed
to atrt and the script itself, to bring about a clear understanding between
them.

Currently autotest-run.sh script accepts arguments whose names are such
that it is difficult to differentiate if the argument is being passed to
atrt or used by the script itself also making it difficult when changes
are to be made to them.

This change ensures there is a clear distinction between the set of
arguments passed to atrt and the script itself. Arguments that are used
only by atrt are now passed as '--atrt-ARGUMENT' and the remaining
arguments do not have any name changes.

Change-Id: I8ab322369789ff3fff0f437508a7ceb128da36ca