Bug#32248291: ENABLE THE HYPERGRAPH OPTIMIZER FOR SQL_BUFFER_RESULT [3/3, enable]
Enable use of the hypergraph optimizer for queries that have the
SQL_BUFFER_RESULT option set. When this option is set, the final query
result is stored in a temporary table before it is returned to the
client.

Supporting SQL_BUFFER_RESULT in the hypergraph optimizer is a
precondition for supporting INSERT INTO ... SELECT, since INSERT INTO
... SELECT enables SQL_BUFFER_RESULT implicitly when the target table
is also accessed by the SELECT statement.

Change-Id: I5da38206e03701b47f03407d407d5302de186935