WL#14476 Introduce NdbScanFilter option for NULL-compare according to SQL semantics
Introduce 'NullSemantics' in the Interpreter instructions:

 -  BRANCH_ATTR_OP_ARG (Compare 'Attr' with constant 'Arg').
 -  BRANCH_ATTR_OP_ATTR (Compare two 'Attr')
 -  BRANCH_ATTR_OP_ARG_2 (Compare 'Attr' with 'ParamNo')

The two spare bits 6 & 7 in the above interpreter opcodes are used to specify the
NULL-values compare semantic to be used for the opcodes. These bits will be
both '0' prior to this patch, which will then map to the NullSemantics 'NULL_CMP_EQUAL'
which will behave as prior to this patch. That is, 'NULL == NULL' and 'NULL < x'.

The other two NullSemantics modes are:

 - IF_NULL_BREAK_OUT: Used when the compare op is part of an AND-GROUP
   If a NULL value is compared, the condition should evaluate to
   'unknown'. As the interpreter only implement two-valued boolean condition (true, or false)
   'unknown' is 'not true' -> 'false'. Thus we can conclude the AND-GROUP is 'not true'
   as well, and the 'break out' jump should be taken.
 - IF_NULL_CONTINUE: Used when the compare op is part of an OR-GROUP.
   If a NULL value is compared, the condition itself becomes 'unknown'. However,
   other conditions in the same OR-group may evaluate to 'true', so we should
   continue with next opcode.

For both of the above modes, compare of the operand values itself is skipped.

Patch also remove the unused 'Array length diff' and 'Varchar flag' from the
BRANCH_ATTR_OP_ARG opcode.

Reviewed By: Martin Skold <Martin.Skold@oracle.com>