WL#14476 Introduce NdbScanFilter option for NULL-compare according to SQL semantics
More refactoring of NdbScanFilter:

Patch introduce the two 'negate tables' negateBinary[] and negateUnary[]
which may be used to find the negated form of the 'compare op'.

In combination with the previous patch, which eliminated (dead code) for
NAND and NOR group, we use these negate tables to prove that the massive
branch tables (table1[] .. table4[]) may be reduced. We can observe that:

For a certain 'compare op', the branch tables will contain the same
branch function as for its negated compare op if we also invert the
AND/OR group.
Thus, we genarally have.

 - table[op].m_branches[AND] == table[negatedOp].m_branches[OR]
 - table[op].m_branches[OR]  == table[negatedOp].m_branches[AND]

This property is used by this patch such that we will only use
the branch function specified for the OR group. The negate tables
will be used in combination with the rules above to achieve this.

This is a pure preparation for the next patch...

Reviewed By: Martin Skold <Martin.Skold@oracle.com>