Bug #32365486: USE-AFTER-FREE WITH HYPERGRAPH OPTIMIZER + SORT + INFORMATION SCHEMA TABLES
If an information schema table is on the right side of a nested loop
join, it would get deleted and re-filled many times. This is not only
inefficient, but if there is a sort by row ID above it (which can only
happen in the hypergraph optimizer; the old optimizer materializes
joins before sorting them), it could end up reading rows that no longer
existed.

Fix by simply not re-filling the table if the data is already in place.
In the process, clean out some unused code and flags related to the
filling.

Change-Id: Ib43e34d7362ee409fb3bed8e11a4a7963f926870