Bug#32291841: Trigger error in create_table_precheck when creating temporary table
The problem occurs when attempting to prepare a query expression
for use as a new temporary table, in repeated executions, such as
within a trigger invocation.

Just before the preparation, the TABLE_LIST object for the new table
is removed temporarily from the SQL command object, and the intention
is to restore the object after the preparation, so that the command
object can be reused later. However, if the preparation causes an
error, the object is not restored, and in next execution a damaged
command object is used.

The fix is to restore the table object in all cases of error.

Reviewed by: Catalin Besleaga <catalin.besleaga@oracle.com>