Bug#32316358 - AUTOTEST: TESTNODERESTART -N LCP_WITH_MANY_PARTS/_DROP_TABLE -R 1000 T17
In order to change a config variable of a data node, the test calls
setConfigValueAndRestartNode(). This method changes the config on
mgmd. It then restarts it by calling restartOneDbNode() and then wait
for it to start (waitNodesStarted()), instead of restarting the data
nodes.

However, the test first calls setConfigValueAndRestartNode() for the
mgmd node (id=1). Since it is an mgmd node, waitNodesStarted() times out.

This fix
- extracts relevant fixes from 8.0 and backport it to 7.6.
- changes setConfigValueAndRestartNode() to return the old config value.
- uses it to change the config values of the *data* nodes only (not mgmd).

Reviewed by: Priyanka Sangam <priyanka.sangam@oracle.com>