Bug#32335496 CONTRIBUTION BY FACEBOOK:   FIX REPLICA THREADS NOT SHOWING IN VARIABLES_BY_THREAD
Fix based on a contribution by FACEBOOK.

Before this fix, replication threads running in the server were:
- visible in table performance_schema.threads (ok)
- not visible in tables performance_schema.variables_by_thread (bug)
- not visible in tables performance_schema.status_by_thread (bug)

The root cause is a missing association between:
- the performance schema thread instrumentation
- the SQL layer THD object
which is required for the performance schema to
inspect the THD object attributes.

The fix is to call mysql_thread_set_psi_THD()
to properly instrument the replication threads.

Approved by: Chris Powers <chris.powers@oracle.com>