Bug#32224707 CPU 100% MYSQL 8 ON WINDOWS 2019
Issue
=====
1. On windows, running SELECT COUNT(*) queries parallely from various client
threads, with the number of client threads exceeding the number of cores of the
machine, can cause stalls for a very long time. The stall is seen in
IB_thread::start(), while spawning threads for parallel read, where we spin-wait
in a loop waiting for the state to be set.

It is seen that we cannot just rely on UT_RELAX_CPU() when the number of threads
to be spawned exceeds the number of cores of a machine. This is because
UT_RELAX_CPU uses the YieldProcesser() system call,  which only makes the
processor available to other logical processors. But if all the logical
processors are busy in spin wait then no real work is done causing the
hang-like, situation.

2. The parallel reader infrastructure, allows the user to set the parameter
innodb_parallel_reader_threads (default being 4), to configure the number of
threads to be spawned to carry out the parallel read operation. This parameter
is supposed to be an upper limit, with the infrastructure using only as many
threads required as possible. But we currently spawn as many threads as the
parameter value even though it's not necessary.

Fix
===
1. If we are still stuck in spin-wait, after calling UT_RELAX_CPU multiple
times, allow the thread to release its time slice.

The patch for this was contributed by Daniel Blanchard
<daniel.blanchard@oracle.com>.

2. Make the parallel reader infrastructure spawn only as many threads as
required, with innodb_parallel_reader_threads as the max limit.

Reviewed-by: Sunny Bains <sunny.bains@oracle.com>
RB: 25645