Bug#32338771 connection stalls network errors to primary
Problem
=======

In case of network-loss between router and group-replication PRIMARY
(but not network loss between group-replication members),
the wait for the failover by the router fails to timeout after 10
seconds.

The wait_primary_failover() method was calculating the "timeout_left"
incorrectly which lead to increasing timeout-left, instead of
descreasing it.

Change
======

- fixed the calculation of the end-time of the timeout.

RB: 25769