Bug#32367019 MySQL server problem in optimizer
This is a problem with the Item_cache object, when the cached_field
member is used to point at a table field whose value is cached
by the object.
When this Item_cache object is used in a prepared statement or
procedure, it may point to a Field object that was used in a prior
execution.

The solution is to replace the member with an Item_field, which is
persistent for the lifetime of the procedure, and whose field member
is properly maintained by Item_field::bind_fields().

Reviewed by: Catalin Besleaga <catalin.besleaga@oracle.com>