Bug #32321201 WHEN SERVER REACHED MAX CONNECTION, ROUTER PENALTY INCOMING CLIENT
The problem is that the current Router implementation does not expect
Error message from the server as a first packet. It only expects Greetings
message. As a result when max connections limit is reached on a server side,
Router fails to decode the Error from the server, sends some generic error
to the Client and also treats the client connection attempt as a connection
error (incrementing the error counter for the client incoming address).

This patch addresses those issues:
1. It forwards the server error to the Client when there is an error in
place of a Greeting message.
2. The connections error counter does not get incremented.

RB: 25766