WL #14325: Interesting orders for hypergraph join optimizer [patch 8/22, unique indexes]
Collect functional dependencies from unique indexes; e.g. if (a,b) is a
unique index for a table, then add FDs {a,b} → c for all other fields c.
These FDs always hold, as opposed to join conditions which only hold
after the join is done. Thus, we extend the NFSM-to-DFSM conversion so
that it always follows such FDs (they are “for free”), and they do not
need to be tracked with their own bitmaps. This reduces the number of
states by a fair amount.

Change-Id: Id42aa9f071b02bbe8eb2df82cb8637f855e52474