Bug#32393421 network loss leads to infinitely blocked primary
Problem
=======

When the router fails to the connect to the PRIMARY it waits 10 seconds
for a new PRIMARY to be elected.

That wait-for-new-primary depends on:

1. the PRIMARY being marked as unreachable
2. the metadata-cache to remove the "unreachable" due to a change of the
   GR information.

That assumption works if the reason for the connection-failure is a
death of the primary and GR electing a new primary.

But if only the network between router and PRIMARY is lost, then the
GR membership information will not change and the nodes availability
will not be reset from "Unavailable" to "ReadWrite".

Change
======

- Don't rely on the instance's availability logic for
  wait-primary-failover anymore. Instead, check the metadata-cache's
  replicaset info directly for a new PRIMARY.
- Don't wait for failover after a connect-timeout

RB: 25771