Bug#31496943: INITIALIZE WITH INNODB_PAGE_SIZE=4096 GETS "SPECIFIED KEY WAS TOO LONG" ERRORS
Problem: Initializing a database with 4K page size would print error
messages in the error log.

Solution: Install a proper error_handler_hook before creating the DD
tables, as this invokes the error handler added to suppress the "key too
long" errors, (the error hook normally installed during initialize does
not do this, and instead prints every error to stderr).

In addition the error_handler_hook was made into a std::atomic, since it
is read and modified from multiple threads. To simplify the code the
type alias ErrorHandlerFunctionPointer was created for the function pointer
type.

Removed fatal_error_handler_hook since it was not used for anything
(other than as an alias for my_message_stderr() in my_malloc).

Change-Id: I971b081eeb9832e9789bf81928862bef9139a87d